refactor(admin): move admin controller to admin namespace
